requirements/dev.txt
-r base.txt

asgiref==3.8.1
django-environ==0.12.0
sqlparse==0.5.3
typing_extensions==4.14.0